+++ src/tasks/vim.rs
@@ -41,7 +41,7 @@ pub fn env(mut exports: Exports) -> Exports {
 
 pub fn task() -> Task {
     Task {
-        name: "vim".to_string(),
+        name: String::from("vim"),
         sync,
         update,
     }