+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - golang: Windows: maybe fix bad download URL (#190)
 
+- better error handling in GitHub Release tasks (#191)
+
 ## [0.27.0] - 2019-02-02
 
 ### Added